feat: separate signer int test into a separate module. Refactor batchSigner tests.
Refs: XRDDEV-2897
