--- src/common/common-int-test/src/main/java/org/niis/xroad/common/test/signer/hook/SignerProxyAfterSuiteHook.java
@@ -1,58 +1,0 @@
-/*
- * The MIT License
- * Copyright (c) 2019- Nordic Institute for Interoperability Solutions (NIIS)
- * Copyright (c) 2018 Estonian Information System Authority (RIA),
- * Nordic Institute for Interoperability Solutions (NIIS), Population Register Centre (VRK)
- * Copyright (c) 2015-2017 Estonian Information System Authority (RIA), Population Register Centre (VRK)
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-
-package org.niis.xroad.common.test.signer.hook;
-
-import com.nortal.test.core.services.hooks.AfterSuiteHook;
-import com.nortal.test.testcontainers.TestableApplicationContainerProvider;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-@Slf4j
-@Component
-@ConditionalOnProperty(value = "test-automation.custom.signer-container-enabled", havingValue = "true")
-@RequiredArgsConstructor
-public class SignerProxyAfterSuiteHook implements AfterSuiteHook {
-
-    private final TestableApplicationContainerProvider containerProvider;
-
-    @Override
-    public void afterSuite() {
-        log.info("Setting permissions for signer files so they could be deleted");
-        try {
-            containerProvider.getContainer().execInContainer("chmod", "-R", "777", "/etc/xroad/signer/");
-        } catch (IOException e) {
-            log.error("Failed to change file permissions", e);
-        } catch (InterruptedException e) {
-            Thread.currentThread().interrupt();
-        }
-    }
-
-}
