+++ src/lib/keyconf-impl/src/main/java/org/niis/xroad/keyconf/impl/KeyConfImpl.java
@@ -62,7 +62,7 @@ import java.util.List;
  */
 @Slf4j
 @RequiredArgsConstructor
-class KeyConfImpl implements KeyConfProvider {
+public class KeyConfImpl implements KeyConfProvider {
     protected final GlobalConfProvider globalConfProvider;
     protected final ServerConfProvider serverConfProvider;
     protected final SignerRpcClient signerRpcClient;
