+++ src/service/proxy/proxy-core/build.gradle.kts
@@ -1,15 +1,6 @@
 plugins {
   id("xroad.java-conventions")
-  id("xroad.int-test-conventions")
   id("xroad.test-fixtures-conventions")
-}
-
-sourceSets {
-  named("intTest") {
-    resources {
-      srcDir("../../../common/common-int-test/src/main/resources/")
-    }
-  }
 }
 
 dependencies {
@@ -47,44 +38,6 @@ dependencies {
   testFixturesImplementation(testFixtures(project(":lib:keyconf-impl")))
   testFixturesImplementation(testFixtures(project(":lib:serverconf-impl")))
   testFixturesImplementation(libs.wsdl4j)
-
-  "intTestRuntimeOnly"(project(":service:signer:signer-application"))
-  "intTestImplementation"(project(":common:common-test"))
-  "intTestImplementation"(project(":common:common-int-test"))
-}
-
-tasks.register<Test>("intTest") {
-  useJUnitPlatform()
-  dependsOn(":service:signer:signer-application:shadowJar")
-
-  description = "Runs integration tests."
-  group = "verification"
-
-  testClassesDirs = sourceSets["intTest"].output.classesDirs
-  classpath = sourceSets["intTest"].runtimeClasspath
-
-  val intTestArgs = mutableListOf<String>()
-
-  if (project.hasProperty("intTestProfilesInclude")) {
-    intTestArgs += "-Dspring.profiles.include=${project.property("intTestProfilesInclude")}"
-  }
-
-  jvmArgs(intTestArgs)
-
-  testLogging {
-    showStackTraces = true
-    showExceptions = true
-    showCauses = true
-    showStandardStreams = true
-  }
-
-  reports {
-    junitXml.required.set(false)
-  }
-}
-
-tasks.named("check") {
-  dependsOn(tasks.named("intTest"))
 }
 
 tasks.withType<Test> {
