--- src/service/proxy/proxy-core/src/intTest/resources/application-override.yml
@@ -1,38 +1,0 @@
----
-#Logging configuration
-logging:
-  level:
-    ROOT: INFO
-    cucumber: TRACE
-    liquibase: WARN
-    org.springframework: INFO
-    org.niis: TRACE
-    com.nortal.test: INFO # TRACE is helpful for development
-    org.niis.xroad.proxy.core.signature.BatchSigner: INFO
-
-test-automation:
-  report-name: xroad-proxy-test-suite
-  spring-component-scan: "org.niis.xroad.common.test,org.niis.xroad.proxy.test"
-  cucumber:
-    execution:
-      parallel:
-        enabled: false
-    glue-append: "org.niis.xroad.common.test.glue,org.niis.xroad.proxy.test.glue"
-    filter:
-      tags: "not @Skip"
-  containers:
-    testable-container:
-      reuse-between-runs: ${reuse-between-runs}
-      directory-mounts:
-        - "/tmp/xroad/passwordstore/:build/container-passwordstore/"
-    context-containers:
-      ca-server:
-        enabled: true
-        relative-a2c-path: "../../../../development/acme2certifier"
-        reuse-between-runs: ${reuse-between-runs}
-  custom:
-    signer-container-enabled: true
-# toggle for reusable containers. This allows quicker test development as containers are not destroyed between runs.
-# WARNING: this leaves containers running indefinitely. They have to be stopped manually.
-# Note: this required testcontainers.reuse.enable=true property to be defined in your ~/.testcontainers.properties file
-reuse-between-runs: false
