--- src/service/signer/signer-core/src/intTest/java/org/niis/xroad/signer/test/container/ContainerSetup.java
@@ -1,56 +1,0 @@
-/*
- * The MIT License
- * Copyright (c) 2019- Nordic Institute for Interoperability Solutions (NIIS)
- * Copyright (c) 2018 Estonian Information System Authority (RIA),
- * Nordic Institute for Interoperability Solutions (NIIS), Population Register Centre (VRK)
- * Copyright (c) 2015-2017 Estonian Information System Authority (RIA), Population Register Centre (VRK)
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package org.niis.xroad.signer.test.container;
-
-import com.nortal.test.testcontainers.configuration.TestableContainerProperties;
-import com.nortal.test.testcontainers.configurator.TestContainerConfigurator;
-import lombok.extern.slf4j.Slf4j;
-import org.niis.xroad.common.test.signer.container.BaseTestSignerSetup;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Slf4j
-@Configuration
-public class ContainerSetup extends BaseTestSignerSetup {
-
-    @Bean
-    public TestContainerConfigurator testContainerConfigurator(
-            TestableContainerProperties testableContainerProperties) {
-        return super.testContainerConfigurator(testableContainerProperties,
-                "../signer-application/build/libs/signer-1.0.jar",
-                "../../../addons/hwtoken/build/libs/hwtoken-1.0.jar");
-    }
-
-    @Bean
-    public TestContainerConfigurator.TestContainerInitListener testContainerInitListener() {
-        return super.testContainerInitListener(true);
-    }
-
-    @Override
-    public String getPksWrapperDir() {
-        return "../../../libs/pkcs11wrapper/%s/%s";
-    }
-}
