--- src/service/signer/signer-core/src/intTest/resources/application-ci.yml
@@ -1,3 +1,0 @@
-test-automation:
-  custom:
-    grpc-client-host-override:  "host.docker.internal"
