+++ src/service/signer/signer-int-test/src/intTest/java/org/niis/xroad/signer/test/SignerClientHolder.java
