+++ src/service/signer/signer-int-test/src/intTest/resources/behavior/0100-signer-software-token.feature
