+++ src/service/signer/signer-int-test/src/intTest/resources/signer-container-files/Dockerfile
