+++ src/settings.gradle.kts
@@ -63,6 +63,7 @@ include("service:signer:signer-api")
 include("service:signer:signer-core")
 include("service:signer:signer-cli")
 include("service:signer:signer-client")
+include("service:signer:signer-int-test")
 
 include("service:message-log-archiver:message-log-archiver-application")
